requirements.txt
alembic==1.7.5
argon2-cffi==21.1.0
attrs==21.2.0
black==21.11b1
cffi==1.15.0
click==8.0.3
Faker==9.9.0
Flask==2.0.2
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
mypy-extensions==0.4.3
mysqlclient==2.1.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
regex==2021.11.10
six==1.16.0
SQLAlchemy==1.4.27
text-unidecode==1.3
toml==0.10.2
tomli==1.2.2
typing_extensions==4.0.1
Werkzeug==2.0.2
